win pc

+++ springbootlearning-master/SpringBoot_studend-crud-restapi-simple/src/main/java/studend_crud_restapi_simple/SpringBoot_studend_crud_restapi_simple/exception/ResourceNotFoundException.java
@@ -3,7 +3,7 @@ package studend_crud_restapi_simple.SpringBoot_studend_crud_restapi_simple.excep
 public class ResourceNotFoundException extends RuntimeException{
     private String resourceName ;
     private String fieldName ;
-    private long fieldValue;
+    private Integer fieldValue;
 
 
     public String getResourceName() {
@@ -16,11 +16,11 @@ public class ResourceNotFoundException extends RuntimeException{
         return fieldName;
     }
 
-    public long getFieldValue() {
+    public Integer getFieldValue() {
         return fieldValue;
     }
 
-    public ResourceNotFoundException (String resourceName , String fieldName , long fieldValue){
+    public ResourceNotFoundException (String resourceName , String fieldName , Integer fieldValue){
         super(String.format("%s not found with %s : %s",resourceName,fieldName,fieldValue));
         this.resourceName = resourceName ;
         this.fieldName = fieldName ;
